+++ ims_python_helper/__init__.py
@@ -32,7 +32,7 @@ from datetime import datetime
 
 import sys
 from time import sleep
-from typing import Dict, List, Optional
+from typing import Dict, List
 
 from pkg_resources import get_distribution
 
@@ -58,6 +58,16 @@ DEFAULT_IMS_API_URL = 'https://api-gw-service-nmn.local/apis/ims'
 
 class ImsImageAlreadyUploaded(Exception):
     """A populated image with some given name already exists in IMS."""
+
+    def __init__(self, image_record: Dict, *args):
+        super().__init__(*args)
+        self.image_record = image_record
+
+    def __str__(self):
+        return "image with name \"{}\" already exists in IMS (ID: {})".format(
+            self.image_record.get("name"),
+            self.image_record.get("id"),
+        )
 
 
 class ImsHelper(object):
@@ -250,43 +260,49 @@ class ImsHelper(object):
         resp.raise_for_status()
         return resp
 
-    def get_empty_image_record_for_name(self, image_name: str) -> Dict:
+    def get_empty_image_record_for_name(self, image_name: str, skip_existing: bool) -> Dict:
         """Get an empty record for an image in IMS with the given name.
 
-        If an image with the given name does not exist, and empty image
+        If an image with the given name does not exist, an empty image
         record will be created. If an empty image record exists, it will
-        be returned. If an uploaded image record exists, then ImsImageExists
-        will be raised.
+        be returned. If an uploaded image record exists, then
+        ImsImageAlreadyUploaded will be raised with the given image record
+        if `skip_existing` is True, and a new image will be created and
+        returned if `skip_existing` is False.
 
         Args:
             image_name: the desired name of the image
-
-        Raises:
-            ImsImageExists: if the image already exists and has been uploaded
+            skip_existing: if True, check if an image exists with the name
+                `image_name`. If it exists, check if it has a `link` attribute
 
         Returns:
             dict: the empty IMS image record for the image, containing e.g.
                 the name, the id, and the creation timestamp
-        """
-        try:
-            existing_images = self._ims_images_get()
-            for image in existing_images:
-                if image.get("name") == image_name:
-                    if image.get("link"):
-                        raise ImsImageAlreadyUploaded(image_name)
-                    else:
-                        LOGGER.info("Image \"%s\" with ID \"%s\" found, but it has not been "
-                                    "uploaded yet; uploading now.", image_name, image.get("id"))
-                        return image
-        except requests.HTTPError as err:
-            LOGGER.warning("Could not retrieve existing images: %s", err)
 
-        LOGGER.info("No image found in IMS with name \"%s\"; creating.", image_name)
+        Raises:
+            ImsImageAlreadyUploaded: if the image already exists and
+                has been uploaded, and `skip_existing` is True
+        """
+        if skip_existing:
+            try:
+                existing_images = self._ims_images_get()
+                for image in existing_images:
+                    if image.get("name") == image_name:
+                        if image.get("link"):
+                            raise ImsImageAlreadyUploaded(image)
+                        else:
+                            LOGGER.info("Image \"%s\" with ID \"%s\" found, but it has not been "
+                                        "uploaded yet.", image_name, image.get("id"))
+                            return image
+            except requests.HTTPError as err:
+                LOGGER.warning("Could not retrieve existing images: %s", err)
+
+        LOGGER.info("Creating image with name \"%s\"", image_name)
         return self._ims_image_create(image_name)
 
     def image_upload_artifacts(
             self, image_name, ims_job_id=None, rootfs=None, kernel=None,
-            initrd=None, debug=None, boot_parameters=None
+            initrd=None, debug=None, boot_parameters=None, skip_existing=False,
     ):
         """
         Utility function to upload and register any image artifacts with the
@@ -294,8 +310,9 @@ class ImsHelper(object):
         values are expected to be full paths to readable files. Only squashfs
         is currently supported for the rootfs parameter.
 
-        If an image with name `image_name` and a `link` attribute already
-        exists in IMS, then this function will return without uploading anything.
+        If `skip_existing` is True and an image with name `image_name` and a
+        `link` attribute already exists in IMS, then this function will return
+        the existing image record without uploading anything.
         """
 
         # Stub out the return value of this method
@@ -305,10 +322,11 @@ class ImsHelper(object):
         }
 
         try:
-            image_record = self.get_empty_image_record_for_name(image_name)
-        except ImsImageAlreadyUploaded:
+            image_record = self.get_empty_image_record_for_name(image_name, skip_existing)
+        except ImsImageAlreadyUploaded as exc:
             LOGGER.warning("Image with name %s already exists in IMS; skipping.", image_name)
-            return
+            ret['ims_image_record'] = exc.image_record
+            return ret
 
         ret["ims_image_record"] = image_record
         image_id = ret["ims_image_record"]["id"]
@@ -327,6 +345,9 @@ class ImsHelper(object):
         if boot_parameters:
             to_upload.append(('application/vnd.cray.image.parameters.boot', key % 'boot_parameters',
                               boot_parameters[0]))  # noqa: E501
+        if not to_upload:
+            LOGGER.info("No image artifacts supplied for image %s; not uploading anything.", image_name)
+            return ret
 
         # Upload the files in series
         LOGGER.info(